Extract colony validation to ColonyValidator.
